🚚 Move response schemas to own folder
